+++ src/librapid/utils/array_utils.hpp
@@ -60,7 +60,7 @@ namespace librapid
 			{
 				throw std::invalid_argument("CUDA support was not enabled, so an"
 											" Array on the GPU cannot be printed");
-		}
+			}
 		#endif
 
 			std::string str = stream.str();
@@ -70,7 +70,7 @@ namespace librapid
 				return;
 			}
 
-			size_t index = str.find_last_of('.');
+			int64_t index = str.find_last_of('.');
 			if (index == std::string::npos)
 			{
 				res = {str.length(), 0};
@@ -78,7 +78,7 @@ namespace librapid
 			}
 
 			res = {index, str.length() - index - 1};
-	}
+		}
 
 		inline void autocastFormatValue(const RawArray &src, std::string &res)
 		{
@@ -135,7 +135,7 @@ namespace librapid
 			res = stream.str();
 			if (isFloating(src.dtype) && res.find_last_of('.') == std::string::npos)
 				res += ".";
-			}
+		}
 
 		template<typename _Ty>
 		void arrayOpEq(void *dataStart, Accelerator location, const _Ty &val)
@@ -158,7 +158,7 @@ namespace librapid
 			#endif // LIBRAPID_HAS_CUDA
 			}
 		}
-		}
+	}
 }
 
 #endif // LIBRAPID_ARRAY_UTILS
